+++ src/main/java/abuteen/hashem/ordermanagementbackend/entity/ProductOrder.java
@@ -26,8 +26,8 @@ public class ProductOrder implements Serializable {
     private Order order;
 
     private int quantity;
-    private float price;
-    private float vat;
+    private double price;
+    private double vat;
