+++ src/main/java/abuteen/hashem/ordermanagementbackend/repository/ProductOrderRepository.java
@@ -3,5 +3,5 @@ package abuteen.hashem.ordermanagementbackend.repository;
 import abuteen.hashem.ordermanagementbackend.entity.ProductOrder;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductOrderRepository extends JpaRepository<ProductOrder, Long> {
+public interface ProductOrderRepository extends JpaRepository<ProductOrder, Integer> {
 }
