+++ src/main/java/abuteen/hashem/ordermanagementbackend/repository/StockRepository.java
@@ -3,5 +3,5 @@ package abuteen.hashem.ordermanagementbackend.repository;
 import abuteen.hashem.ordermanagementbackend.entity.Stock;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StockRepository extends JpaRepository<Stock, Long> {
+public interface StockRepository extends JpaRepository<Stock, Integer> {
 }
